Experiment: readonly and value attributes
input.pet.php is abstracting the readonly attribute by generating a
hidden element an a disabled element when readonly is not supported.

Also, when passed a value attribute it will be converted into checked
for checkboxes and radio.

This means that there is a disparity between the server side view and
the client side. This may cause confusion among developers, but will
facilitate generating those fields programatically for a future form
generator.