Fixed a path resolution issuse in Paladio
Paladio did sometimes mix "/" and "\", this has been fixed.
Also, Parser is now required on top of the file and assumed present.

---

Note: the current architecture of Paladio requires the users to avoid
declaring functions in global scope. Doing that will trigger a faltal
error because Paladio forces the files to be parsed twice, and functions
get declared twice. It is technically possible to fix this, but the
solution is to parse the PHP files from PHP... that would make their
execution slower and the architecture more error prone.