Huge improvements for relative path resolution
Fixed some bugs regarding folders with empty string as name and combined
"\" and "/".
Also added the new convenience function RebaseRelativePath, this
function is equivalent to calling ResolveRelativeFunction and then
CreateRelativeFunction. A function with similar purpose did exist long
ago [before Paladio was public]. It is good to have this back again.