Removed ENTS
ENTS were a way to declare entities in such way that Paladio was able to
create them in the database. At the moment this was introduced it pushed
the evolution of the code.... but this has never been used beyond
initial testing.

If this functionality will come back it will be as a plugin.